// src/features/calendar/types.ts
import type { Dayjs } from "dayjs";

export type EventCategory =
  | 'accounting'
  | 'payroll'
  | 'lodging'
  | 'info'
  | 'other';

export type EventColor =
  | 'red'
  | 'orange'
  | 'blue'
  | 'green'
  | 'primary'
  | 'secondary';

export interface InfoEvent {
  id: string;
  groupId?: string;
  title: string;
  startDate: Date;
  endDate: Date;

  category: EventCategory;
  color: EventColor;
  isAutoGenerated: boolean;

  description?: string;

  employeeIds: string[];
  constructionIds: string[];
}

export interface CalendarEvent {
  id?: string;
  groupId?: string;
  date: Dayjs;
  startDate: Dayjs;
  endDate: Dayjs;
  title: string;
  color: EventColor;
  category: EventCategory;
  description?: string;
  isAutoGenerated?: boolean;
  employeeIds: string[];
  constructionIds: string[];
}

export interface CalendarGridProps {
  onMoreClick: (data: CalendarDay) => void;

  monthGrid: CalendarDay[][];
  currentMonth: Dayjs;
  selectDay: Dayjs | null;
  onDayClick: (day: Dayjs) => void;
  isDayInRange: (day: Dayjs) => boolean;
  handleEventClick: (event: CalendarEvent) => void;
}

export interface CalendarDay {
  date: Dayjs;
  events: CalendarEvent[];
  slots?: Record<string, number>;
}